+++ rs/earthbucks_lib/src/tx_out_bn.rs
@@ -3,5 +3,5 @@ use crate::tx_out::TxOut;
 #[derive(Debug, Clone, PartialEq, Default)]
 pub struct TxOutBn {
     pub tx_out: TxOut,
-    pub block_num: u64,
+    pub block_num: u32,
 }
